+++ testing/src/main/java/uk/gov/pay/commons/testing/pact/consumers/Pacts.java
@@ -1,6 +1,10 @@
 package uk.gov.pay.commons.testing.pact.consumers;
 
-import java.lang.annotation.*;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Repeatable;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
